Move files for generalising interior modules
