+++ src/proteus/interior/__init__.py
@@ -1,0 +1,7 @@
+from __future__ import annotations
+
+from .wrapper_interior import RunInterior
+
+__all__ = [
+    'RunInterior',
+]
