+++ src/proteus/interior/dummy.py
@@ -1,0 +1,12 @@
+# Dummy interior module
+from __future__ import annotations
+
+import logging
+
+import numpy as np
+
+from proteus.utils.helper import UpdateStatusfile
+
+log = logging.getLogger("fwl."+__name__)
+
+
