+++ src/proteus/interior/spider.py
