+++ src/proteus/interior/wrapper_interior.py
@@ -1,0 +1,11 @@
+# Generic interior wrapper
+from __future__ import annotations
+
+import logging
+
+from proteus.utils.constants import AU
+from proteus.utils.helper import PrintHalfSeparator, UpdateStatusfile, safe_rm
+
+log = logging.getLogger("fwl."+__name__)
+
+
